refactor: split _rate_limit.py into config, store, and factory modules
Decompose the 129-line monolithic rate-limit module into three focused
sub-modules (_rate_limit_config, _rate_limit_store, _rate_limit_factory),
each under 100 lines. The original _rate_limit.py becomes a re-export
hub preserving backward compatibility.
